requirements.txt
astropy
matplotlib
numpy
scikit-image
Pillow